requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.30
redis==5.0.4
python-dotenv==1.0.1
pydantic==2.7.1